Switch application to only have an open button
Problem: Trying to run Oasis inside of an Electron window means that we
have to build our own browser (with a back button, tabs, etc), and
that's a ton of work that we should avoid if possible.

Solution: Turn Oasis-Desktop into a launcher type of thing where it just
has a button that opens your actual web browser instead of trying to
display things in the Electron window.