Patch App-Builder-Lib to allow README.md
Problem: Electron-Builder doesn't include any files called `README.md`
because it wants to keep the output size small. This is great, except
that Oasis displays its own readme and needs this file. When you try to
open Oasis-Desktop and this file isn't found, it crashes.

Solution: I tried everything. For the past 24 hours I've been using the
documentation (https://www.electron.build/configuration/contents.html)
as my bible, but I cannot find a way to include the readme. It seems
that the ignored files cannot be un-ignored, and are hard-coded, so the
only solution is to patch App-Builder-Lib. Unfortunately Oasis Desktop
doesn't directly depend on that module, so it's difficult to modify a
deep dependency, and it seems that our only solution is to use a tool
like Patch-Package. This hacky solution applies a patch that removes
`"README.md"` from the list of files that should be excluded from the
output ASAR.

Fixes: https://github.com/fraction/oasis-desktop/issues/3